added permissions to all api functions

+++ mysite/home/api_views.py
@@ -34,34 +34,31 @@ def item_list(request, format=None):
         items = Item.objects.all()
         serializer = ItemSerializer(items, many=True)
         response = Response(serializer.data)
-        if 'HTTP_AUTHORIZATION' in request.META.keys():
-            if Token.objects.get(key=request.META['HTTP_AUTHORIZATION'][6:]):
-                return response
-        elif Token.objects.get(user=request.user):
-            response['Authorization']=Token.objects.get(user=request.user)
-            return response
-        else:
-            return HttpResponse('Please Request an API Token')
+        return response
 
     elif request.method == 'POST':
-        serializer = ItemSerializer(data=request.data)
-        tags=[]
-        if 'tags' in serializer.initial_data.keys():
-            if isinstance(serializer.initial_data['tags'], str):
-                tags.append(serializer.initial_data['tags'])
-            else:
-                for tag in serializer.initial_data['tags']:
-                    tags.append(tag)
-            del serializer.initial_data['tags']
-        if serializer.is_valid():
-            serializer.save()
-            for tag in tags:
-                new_tag=Tag(item_id=Item.objects.get(item_name=serializer.data['item_name']),tag=tag)
-                new_tag.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_staff:
+            serializer = ItemSerializer(data=request.data)
+            tags=[]
+            if 'tags' in serializer.initial_data.keys():
+                if isinstance(serializer.initial_data['tags'], str):
+                    tags.append(serializer.initial_data['tags'])
+                else:
+                    for tag in serializer.initial_data['tags']:
+                        tags.append(tag)
+                del serializer.initial_data['tags']
+            if serializer.is_valid():
+                serializer.save()
+                for tag in tags:
+                    new_tag=Tag(item_id=Item.objects.get(item_name=serializer.data['item_name']),tag=tag)
+                    new_tag.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response('Manager Permission Required')
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes((TokenAuthentication,SessionAuthentication))
+@permission_classes((IsAuthenticated,))
 def item_detail(request, pk, format=None):
     """
     Retrieve, update or delete a snippet instance.
@@ -76,23 +73,49 @@ def item_detail(request, pk, format=None):
         return Response(serializer.data)
 
     elif request.method == 'PUT':
-        serializer = ItemSerializer(item, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_staff:
+            serializer = ItemSerializer(item, data=request.data)
+            tags=[]
+            if 'tags' in serializer.initial_data.keys():
+                if isinstance(serializer.initial_data['tags'], str):
+                    tags.append(serializer.initial_data['tags'])
+                else:
+                    for tag in serializer.initial_data['tags']:
+                        tags.append(tag)
+                del serializer.initial_data['tags']
+            if serializer.is_valid():
+                serializer.save()
+                for tag in Tag.objects.filter(item_id=Item.objects.get(item_name=serializer.data['item_name'])):
+                    if tag.tag not in tags:
+                        tag.delete()
+                for tag in tags:
+                    try: 
+                        Tag.objects.get(item_id=Item.objects.get(item_name=serializer.data['item_name']),tag=tag)
+                    except Tag.DoesNotExist:
+                        new_tag=Tag(item_id=Item.objects.get(item_name=serializer.data['item_name']),tag=tag)
+                        new_tag.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response('Manager Permission Required')
 
     elif request.method == 'DELETE':
-        item.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.is_superuser:
+            item.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response('Administrator Permission Required')
     
 @api_view(['GET', 'POST'])
+@authentication_classes((TokenAuthentication,SessionAuthentication))
+@permission_classes((IsAuthenticated,))
 def request_list(request, format=None):
     """
     List all code snippets, or create a new snippet.
     """
     if request.method == 'GET':
-        requests = Request.objects.all()
+        if request.user.is_staff:
+            requests = Request.objects.all()
+        else:
+            requests = Request.objects.filter(owner=request.user)
         serializer = RequestSerializer(requests, many=True)
         return Response(serializer.data)
 
@@ -104,6 +127,8 @@ def request_list(request, format=None):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes((TokenAuthentication,SessionAuthentication))
+@permission_classes((IsAuthenticated,))
 def request_detail(request, pk, format=None):
     """
     Retrieve, update or delete a snippet instance.
@@ -114,65 +139,82 @@ def request_detail(request, pk, format=None):
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'GET':
-        serializer = RequestSerializer(request_object)
-        return Response(serializer.data)
+        if request_object.owner == request.user or request.user.is_staff:
+            serializer = RequestSerializer(request_object)
+            return Response(serializer.data)
+        return Response('Manager Permission Required')
 
     elif request.method == 'PUT':
-        serializer = RequestSerializer(request_object, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request_object.owner == request.user or request.user.is_staff:
+            serializer = RequestSerializer(request_object, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response('Manager Permission Required')
 
     elif request.method == 'DELETE':
-        request_object.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request_object.owner == request.user:
+            request_object.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response('Owner Permission Required')
 
 @api_view(['GET'])
+@authentication_classes((TokenAuthentication,SessionAuthentication))
+@permission_classes((IsAuthenticated,))
 def user_list(request, format=None):
     """
     List all code snippets, or create a new snippet.
     """
     if request.method == 'GET':
-        users = User.objects.all()
-        serializer = UserSerializer(users, many=True)
-        return Response(serializer.data)
+        if request.user.is_staff:
+            users = User.objects.all()
+            serializer = UserSerializer(users, many=True)
+            return Response(serializer.data)
+        return Response('Manager Permission Required')
     
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes((TokenAuthentication,SessionAuthentication))
+@permission_classes((IsAuthenticated,))
 def user_detail(request, pk, format=None):
     """
     Retrieve, update or delete a snippet instance.
     """
-    try:
-        user = User.objects.get(pk=pk)
-    except Request.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    if request.method == 'GET':
-        serializer = UserSerializer(user)
-        return Response(serializer.data)
-
-    elif request.method == 'PUT':
-        serializer = UserSerializer(user, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
+    if request.user.is_staff:
+        try:
+            user = User.objects.get(pk=pk)
+        except Request.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+    
+        if request.method == 'GET':
+            serializer = UserSerializer(user)
             return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    elif request.method == 'DELETE':
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+    
+        elif request.method == 'PUT':
+            serializer = UserSerializer(user, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+        elif request.method == 'DELETE':
+            user.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    return Response('Manager Permission Required')
     
 @api_view(['POST'])
+@authentication_classes((TokenAuthentication,SessionAuthentication))
+@permission_classes((IsAuthenticated,))
 def user_create(request, format=None):
     """
     List all code snippets, or create a new snippet.
     """
     if request.method == 'POST':
-        serializer = UserCreateSerializer(data=request.data)
-        if serializer.is_valid():
-            user = User.objects.create_user(username=serializer.data['username'],password=serializer.data['password'])
-            user.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
+        if request.user.is_superuser:
+            serializer = UserCreateSerializer(data=request.data)
+            if serializer.is_valid():
+                user = User.objects.create_user(username=serializer.data['username'],password=serializer.data['password'])
+                user.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response('Administrator Permission Required')
